Conflicts:

	gui/Check.cpp
	objects.cpp
	screen/choose/HighScore.cpp
	text/text.cpp

+++ macros.hpp
@@ -22,7 +22,7 @@
 
 #include <cstdio>
 
-#define LENGTH(x) (sizeof(x) / sizeof(*x))
+#define LENGTH(x) ((int) (sizeof(x) / sizeof(*x)))
 
 #define PRINT_INT(i) printf(#i ": %d\n", (i))
 #define PRINT_POINTER(i) printf(#i ": %p\n", (i))
